test(tui): keep scroll tests with scrollable
Move pure scroll geometry and scrollable-block renderer tests into the shared scrollable widget module, leaving widget-specific tests with their widgets.
