+++ src/console/widgets/scrollable.rs
@@ -47,6 +47,16 @@ pub(crate) const fn effective_offset(content_len: usize, viewport: usize, offset
     if offset > max { max } else { offset }
 }
 
+pub(crate) const fn clamp_scroll_offset(
+    content_len: usize,
+    viewport: usize,
+    offset: &mut u16,
+) -> u16 {
+    let effective = effective_offset(content_len, viewport, *offset);
+    *offset = effective;
+    effective
+}
+
 pub(crate) fn scrollbar_position_for_offset(
     content_length: usize,
     viewport: usize,
@@ -439,10 +449,11 @@ pub(crate) fn render_scrollable_block(
 #[cfg(test)]
 mod tests {
     use super::{
-        cursor_follow_offset, render_vertical_scrollbar_in_area,
-        scrollbar_offset_for_track_position, scrollbar_thumb_geometry,
+        clamp_scroll_offset, cursor_follow_offset, render_scrollable_block,
+        render_vertical_scrollbar_in_area, scrollbar_offset_for_track_position,
+        scrollbar_position_for_offset, scrollbar_thumb_geometry,
     };
-    use ratatui::{Terminal, backend::TestBackend, layout::Rect};
+    use ratatui::{Terminal, backend::TestBackend, layout::Rect, text::Line};
 
     #[test]
     fn scrollbar_thumb_length_is_offset_invariant() {
@@ -492,8 +503,153 @@ mod tests {
     }
 
     #[test]
+    fn clamp_scroll_offset_updates_stored_offset() {
+        let mut scroll_x = 400;
+
+        let effective = clamp_scroll_offset(100, 60, &mut scroll_x);
+
+        assert_eq!(effective, 40);
+        assert_eq!(scroll_x, 40);
+    }
+
+    #[test]
+    fn apply_scroll_delta_moves_from_clamped_offset() {
+        let mut scroll_x = 40;
+
+        super::apply_scroll_delta(&mut scroll_x, -8);
+
+        assert_eq!(scroll_x, 32);
+    }
+
+    #[test]
     fn track_position_maps_to_scrollbar_thumb_range() {
         assert_eq!(scrollbar_offset_for_track_position(20, 5, 10, 0), 0);
         assert_eq!(scrollbar_offset_for_track_position(20, 5, 10, 9), 15);
+    }
+
+    #[test]
+    fn scrollbar_position_maps_visible_end_to_track_end() {
+        assert_eq!(scrollbar_position_for_offset(13, 10, 0), 0);
+        assert_eq!(scrollbar_position_for_offset(13, 10, 3), 3);
+    }
+
+    #[test]
+    fn scrollbar_position_clamps_overscroll() {
+        assert_eq!(scrollbar_position_for_offset(13, 10, 99), 3);
+    }
+
+    #[test]
+    fn scrollable_block_scrollbar_thumbs_reach_visible_ends() {
+        let backend = TestBackend::new(12, 6);
+        let mut terminal = Terminal::new(backend).unwrap();
+        let mut scroll_x = 10;
+        let mut scroll_y = 4;
+        let lines: Vec<Line<'static>> = (0..8)
+            .map(|idx| Line::from(format!("{idx:02}-abcdefghijklmnopq")))
+            .collect();
+
+        terminal
+            .draw(|frame| {
+                render_scrollable_block(
+                    frame,
+                    Rect::new(0, 0, 12, 6),
+                    lines,
+                    &mut scroll_x,
+                    &mut scroll_y,
+                    true,
+                    Some(" Test "),
+                );
+            })
+            .unwrap();
+
+        let buffer = terminal.backend().buffer();
+        assert_eq!(buffer[(10, 5)].symbol(), "━");
+        assert_eq!(buffer[(11, 4)].symbol(), "█");
+    }
+
+    #[test]
+    fn scrollable_block_scrollbar_thumbs_are_proportional_to_viewport() {
+        let backend = TestBackend::new(12, 6);
+        let mut terminal = Terminal::new(backend).unwrap();
+        let mut scroll_x = 0;
+        let mut scroll_y = 0;
+        let lines: Vec<Line<'static>> = (0..5)
+            .map(|idx| Line::from(format!("{idx:02}-abcdefgh")))
+            .collect();
+
+        terminal
+            .draw(|frame| {
+                render_scrollable_block(
+                    frame,
+                    Rect::new(0, 0, 12, 6),
+                    lines,
+                    &mut scroll_x,
+                    &mut scroll_y,
+                    true,
+                    Some(" Test "),
+                );
+            })
+            .unwrap();
+
+        let buffer = terminal.backend().buffer();
+        let horizontal_thumb_len = (1..=10).filter(|x| buffer[(*x, 5)].symbol() == "━").count();
+        let vertical_thumb_len = (1..=4).filter(|y| buffer[(11, *y)].symbol() == "█").count();
+
+        assert_eq!(horizontal_thumb_len, 9);
+        assert_eq!(vertical_thumb_len, 3);
+    }
+
+    #[test]
+    fn scrollable_block_preserves_matching_right_padding_at_horizontal_end() {
+        let backend = TestBackend::new(8, 4);
+        let mut terminal = Terminal::new(backend).unwrap();
+        let mut scroll_x = 99;
+        let mut scroll_y = 0;
+        let lines = vec![Line::from("  abcdefgh")];
+
+        terminal
+            .draw(|frame| {
+                render_scrollable_block(
+                    frame,
+                    Rect::new(0, 0, 8, 4),
+                    lines,
+                    &mut scroll_x,
+                    &mut scroll_y,
+                    true,
+                    Some(" Test "),
+                );
+            })
+            .unwrap();
+
+        let buffer = terminal.backend().buffer();
+        let visible: String = (1..=6).map(|x| buffer[(x, 1)].symbol()).collect();
+
+        assert_eq!(scroll_x, 6);
+        assert_eq!(visible, "efgh  ");
+    }
+
+    #[test]
+    fn scrollable_block_clamps_scroll_y_in_place() {
+        let backend = TestBackend::new(12, 6);
+        let mut terminal = Terminal::new(backend).unwrap();
+        let mut scroll_x = 0;
+        let mut scroll_y = 99;
+        let lines: Vec<Line<'static>> = (0..8).map(|idx| Line::from(format!("{idx:02}"))).collect();
+
+        terminal
+            .draw(|frame| {
+                render_scrollable_block(
+                    frame,
+                    Rect::new(0, 0, 12, 6),
+                    lines,
+                    &mut scroll_x,
+                    &mut scroll_y,
+                    false,
+                    None,
+                );
+            })
+            .unwrap();
+
+        assert_eq!(scroll_y, 4);
     }
 }
